requirements.txt
[[source]]
url = "https://pypi.org/simple"
verify_ssl = true
name = "pypi"

[packages]
streamlit = "*"
langchain-groq = "*"
langchain-community = "*"
pypdf = "*"
sentence-transformers = "*"

[dev-packages]

[requires]
python_version = "3.11"


